+++ tests/testthat/test-algo-seasonal_decomp_arima.R
@@ -1,5 +1,5 @@
 # ---- STANDARD ARIMA ----
-context("TEST seasonal_decomp() - stlm_arima")
+context("TEST seasonal_reg() - stlm_arima")
 
 
 # SETUP ----
@@ -8,15 +8,15 @@ context("TEST seasonal_decomp() - stlm_arima")
 splits <- initial_time_split(taylor_30_min, prop = 0.9)
 
 # Model Spec
-model_spec <- seasonal_decomp(seasonal_period_1 = "1 day", seasonal_period_2 = "week") %>%
+model_spec <- seasonal_reg(seasonal_period_1 = "1 day", seasonal_period_2 = "week") %>%
     set_engine("stlm_arima")
 
 # CHECKS ----
-test_that("seasonal_decomp: checks", {
+test_that("seasonal_reg: checks", {
 
     # external regressors message
     expect_error({
-        seasonal_decomp(seasonal_period_1 = 1) %>%
+        seasonal_reg(seasonal_period_1 = 1) %>%
             set_engine("stlm_arima") %>%
             fit(value ~ date, data = training(splits))
     })
@@ -38,7 +38,7 @@ predictions_tbl <- model_fit %>%
 
 
 # TESTS
-test_that("seasonal_decomp - arima: parnip", {
+test_that("seasonal_reg - arima: parnip", {
 
     testthat::expect_s3_class(model_fit$fit, "stlm_arima_fit_impl")
 
@@ -103,7 +103,7 @@ predictions_tbl <- wflw_fit %>%
 
 
 # TESTS
-test_that("seasonal_decomp: workflow", {
+test_that("seasonal_reg: workflow", {
 
     testthat::expect_s3_class(wflw_fit$fit$fit$fit, "stlm_arima_fit_impl")