+++ NEWS.md
@@ -1,10 +1,23 @@
 
 # modeltime 0.0.2.9000 (Development Version)
 
+### New Models
+
+TBATS Model
+
+``` r
+seasonal_reg() %>%
+    set_engine("tbats")
+```
+
 ### Fixes
 
 - `modeltime_forecast()`: More descriptive errors when external regressors are required. 
 - `prophet_reg()` & `prophet_boost()`: More descriptive error message when less than 100 observations are trained on. Prophet requires >= 100 observations per [facebook/prophet#1554](https://github.com/facebook/prophet/issues/1554)
+
+### Breaking Changes
+
+- `seasonal_decomp()` has changed to `seasonal_reg()`
 
 # modeltime 0.0.2